% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryStatistics.R
\name{as.matrix.QDDictionaryStatistics}
\alias{as.matrix.QDDictionaryStatistics}
\title{as.matrix}
\usage{
\method{as.matrix}{QDDictionaryStatistics}(x, ...)
}
\arguments{
\item{dstat}{A}
}
\description{
Extract the number part from a \code{QDDictionaryStatistics} table as a matrix
}
