% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrama.R
\docType{data}
\name{atypes}
\alias{atypes}
\title{a list of names of the UIMA annotation types used}
\format{An object of class \code{list} of length 20.}
\usage{
atypes
}
\description{
a list of names of the UIMA annotation types used
}
\keyword{datasets}
