% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot.QDCharacterStatistics}
\alias{barplot.QDCharacterStatistics}
\title{Stacked Bar Plot}
\usage{
\method{barplot}{QDCharacterStatistics}(
  height,
  col = qd.colors,
  column = "tokens",
  order = -1,
  labels = TRUE,
  top = 5,
  ...
)
}
\arguments{
\item{height}{The object of class QDCharacterStatistics that is to be plotted}

\item{col}{The colors to use}

\item{column}{Which column of the character statistics should be used?}

\item{order}{Sort the fields inversely}

\item{labels}{Whether to add character labels into the plot}

\item{top}{Limit the labels to the top 5 characters. Otherwise, 
labels will become unreadable.}

\item{...}{All remaining options are passed to \code{barplot.default()}.}
}
\value{
See \code{barplot.default()}.
}
\description{
This function expects an object of type QDCharacterStatistics and 
plots the specified column as a stacked bar plot.
}
\seealso{
barplot.default
}
