% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_dictionary}
\alias{base_dictionary}
\title{Base dictionary}
\format{A list with five entries, each of them being a character vector.}
\usage{
base_dictionary
}
\description{
A list of word fields, i.e., collections of German lemmas associated 
with the five concepts Familie (family), Krieg (war), Liebe (love), Ratio 
(reason) and Religion (religion). The base dictionary is for \emph{demo purposes} 
only, because it doesn't contain any umlaut characters.
}
\keyword{datasets}
