% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{characterMatrix}
\alias{characterMatrix}
\title{Stacked matrix}
\usage{
characterMatrix(fstat, column = "tokens", order = -1)
}
\arguments{
\item{fstat}{The figure statistics table, i.e., the output of figureStatistics(). Coerced to a data.table if needed.}

\item{column}{A column name found in the statistics table. This count is used 
as a basis for the plot.}

\item{order}{If set to -1 (default), figures are ranked descending 
(i.e., figure with most spoken words first). If set to 1, 
figures are ranked ascending.}
}
\description{
This function takes a data frame describing various metrics of figures in dramas 
and creates a matrix that can be used to create a stacked bar plot.
}
