% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{chunkSegmentation}
\alias{chunkSegmentation}
\title{EXPERIMENTAL: Continuous Segmentation}
\usage{
chunkSegmentation(t, n_chunks = 100)
}
\arguments{
\item{t}{Mtext dataframe}

\item{n_chunks}{Number of chunks into which the text will be equally split}
}
\description{
\code{chunkSegmentation} segments the input drama text table into 
chunks of equal size (by token count). Be careful, as this overrides the 
\code{Number.Act}, \code{begin.Act} and \code{end.Act} columns, so that the output 
can be used with act segmentation for functions like \code{dictionaryStatistics}, 
\code{figureStatistics} and \code{frequencytable}.
}
