% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{combine}
\alias{combine}
\title{Combine multiple plays}
\usage{
combine(x, y)
}
\arguments{
\item{y}{A \code{QDDrama}}
}
\value{
A single QDDrama object that represents both plays.
}
\description{
The function \code{combine(x, y)} can be used to merge 
multiple objects of the type \code{QDDrama} into one.
}
\examples{

data(rksp.0)
data(rjmw.0)
d <- combine(rjmw.0, rksp.0)
}
