% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{countAnnotations}
\alias{countAnnotations}
\title{Function to count the annotations of a certain type in selected texts.}
\usage{
countAnnotations(ids, type = "de.unistuttgart.ims.drama.api.Utterance",
  debug = FALSE, shortname = TRUE)
}
\arguments{
\item{ids}{A vector or list of drama ids}

\item{type}{A string, the fully qualified type name we want to count}

\item{debug}{Logical value, whether to print debug information}

\item{shortname}{Logical value, whether to only use the local name of type in the returned data frame.}
}
\description{
Function to count the annotations of a certain type in selected texts.
}
