% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{rksp.0}
\alias{rjmw.0}
\title{Data sets}
\format{A list containing data.frames and data.table.}
\usage{
rksp.0

rjmw.0
}
\description{
\code{rksp.0} represents the data set exported from Emilia Galotti
}
\keyword{datasets}
