% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryStatistics.R
\name{dictionaryStatistics}
\alias{dictionaryStatistics}
\alias{dictionaryStatisticsSingle}
\alias{as.matrix.QDDictionaryStatistics}
\title{Dictionary Use}
\usage{
dictionaryStatistics(drama, fields = base_dictionary[fieldnames],
  fieldnames = c("Liebe"), segment = c("Drama", "Act", "Scene"),
  normalizeByFigure = FALSE, normalizeByField = FALSE,
  byFigure = TRUE, column = "Token.lemma", ci = TRUE)

dictionaryStatisticsSingle(drama, wordfield = c(), segment = c("Drama",
  "Act", "Scene"), normalizeByFigure = FALSE, normalizeByField = FALSE,
  byFigure = TRUE, fieldNormalizer = length(wordfield),
  column = "Token.lemma", ci = TRUE, colnames = NULL)

\method{as.matrix}{QDDictionaryStatistics}(x, ...)
}
\arguments{
\item{fields}{A list of lists that contains the actual field names. 
By default, we load the base_dictionary.}

\item{fieldnames}{A list of names for the dictionaries.}

\item{segment}{The segment level that should be used. By default, 
the entire play will be used. Possible values are "Drama" (default), 
"Act" or "Scene"}

\item{normalizeByFigure}{Logical. Whether to normalize by figure speech length}

\item{normalizeByField}{Logical. Whether to normalize by dictionary size. You usually want this.}

\item{byFigure}{Logical, defaults to TRUE. If false, values will be calculated
for the entire segment (play, act, or scene), and not for individual characters.}

\item{column}{The table column we apply the dictionary on. 
Should be either "Token.surface" or "Token.lemma".}

\item{ci}{Whether to ignore case. Defaults to TRUE, i.e., case is ignored.}

\item{wordfield}{A character vector containing the words or lemmas 
to be counted (only for \code{*Single}-functions)}

\item{fieldNormalizer}{defaults to the length of the wordfield}

\item{colnames}{The column names to be used}

\item{x}{An object of the type \code{QDDictionaryStatistics}, e.g., the output of \code{dictionaryStatistics}.}

\item{t}{A text (data.frame or data.table)}

\item{names}{Logical. Whether the resulting table contains figure ids or names.}
}
\value{
A numeric matrix that contains the frequency with which a dictionary is present in a subset of tokens
}
\description{
These methods retrieve 
count the number of occurrences of the words in the dictionaries, 
across different speakers and/or segments.
The function \code{dictionaryStatistics()} calculates statistics for 
dictionaries with multiple entries, \code{dictionaryStatisticsSingle()} only
for a single word list. Functions ending on \code{L} return a list with 
multiple components.

Extract the number part from a \code{QDDictionaryStatistics} table as a matrix
}
\examples{
# Check multiple dictionary entries
data(rksp.0)
dstat <- dictionaryStatistics(rksp.0, fieldnames=c("Krieg","Familie"))
# Check a single dictionary entries
data(rksp.0)
fstat <- dictionaryStatisticsSingle(rksp.0, wordfield=c("der"))
mat <- as.matrix(dictionaryStatistics(rksp.0, fieldnames=c("Krieg","Familie")))
}
\seealso{
\code{\link{loadFields}} \code{\link{format.QDHasCharacter}}
}
