% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryStatistics.R
\name{dictionaryStatisticsL}
\alias{dictionaryStatisticsL}
\title{Dictionary Statistics as a list}
\usage{
dictionaryStatisticsL(...)
}
\arguments{
\item{...}{All parameters are passed to \code{\link{dictionaryStatistics}}}
}
\description{
This function applies the same logic as \code{\link{dictionaryStatistics}}, but returns the result as a list with multiple components, and the core data as a matrix
}
\details{
The returned list has three named elements:
\describe{
\item{drama}{The drama in which these counts have been counted}
\item{figure}{the figure these values has spoken}
\item{mat}{A matrix containing the actual values}
}
}
\seealso{
\code{\link{dictionaryStatistics}}
}
