% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ensureSuffix}
\alias{ensureSuffix}
\title{Utility functions}
\usage{
ensureSuffix(x, suffix)
}
\arguments{
\item{x}{The character vector}

\item{suffix}{The suffix}
}
\description{
\code{ensureSuffix} makes certain that a character vector ends in 
a given suffix
}
