% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{exploreDramas}
\alias{exploreDramas}
\title{Graphical User Interface}
\usage{
exploreDramas(...)
}
\arguments{
\item{...}{Provide arguments for shiny::runApp()}
}
\value{
Shiny App
}
\description{
The function \code{exploreDramas()} opens a user interface that shows different analyses for the installed dramas in the form of interactive plots.
}
\examples{
exploreDramas(quiet = TRUE)
exploreDramas(launch.browser = TRUE)
}
\seealso{
\code{shiny}
}
