% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{figure.statistics}
\alias{figure.statistics}
\title{This function extracts figure statistics from a drama text table.}
\usage{
figure.statistics(t, names = FALSE, normalize = FALSE)
}
\arguments{
\item{t}{The drama text}

\item{names}{If set to true, the table will contains figure names instead of ids}

\item{normalize}{Normalising the individual columns}
}
\value{
A data frame with the following columns and one row for each figure:
tokens: The number of tokens spoken by that figure
types : The number of different tokens (= types) spoken by each figure
utterances: The number of utterances
utterance_length_mean: The mean length of utterances
utterance_length_sd: The standard deviation in utterance length
}
\description{
This function extracts figure statistics from a drama text table.
}
\examples{
data(rksp.0)
stat <- figure.statistics(rksp.0.text, names = FALSE)
}
