% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{figurematrix}
\alias{figurematrix}
\title{This function takes a data frame describing various metrics of figures in dramas 
and creates a matrix that can be used to create a stacked bar plot.}
\usage{
figurematrix(fstat, column = "tokens")
}
\arguments{
\item{fstat}{The figure statistics table, i.e., the output of figure.statistics()}

\item{column}{A column name found in the statistics table. This count is used 
as a basis for the plot.}
}
\description{
This function takes a data frame describing various metrics of figures in dramas 
and creates a matrix that can be used to create a stacked bar plot.
}
\examples{
data(rksp.0,vndf.0)
text <- rbind(rksp.0.text,vndf.0)
stat <- figure.statistics(text, names = TRUE)
fmatrix <- figurematrix(stat)
# Plot a stacked bar plot
b <- barplot(mat$values,col=qd.colors)
# Add figure names (if needed/wanted)
text(x=b,y=t(mat$cs+(mat$values/2)),labels=t(substr(mat$labels,0,20)))
}
