% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{filter}
\alias{filter}
\title{Filter characters}
\usage{
filter(object, drama, by = c("rank", "tokens"), threshold = ifelse(by
  == "tokens", 500, 10))
}
\arguments{
\item{object}{The object we want to filter}

\item{drama}{The QDDrama object}

\item{by}{Specifies the filter mechanism}

\item{threshold}{The threshold}
}
\description{
This function can be used to filter characters from all tables 
that contain  a character column (and are of the class QDHasCharacter).
}
\examples{
data(rjmw.0)
dstat <- dictionaryStatistics(rjmw.0)
filter(dstat, rjmw.0, by="tokens", threshold=1000)
}
