% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryStatistics.R
\name{filterByDictionary}
\alias{filterByDictionary}
\title{Filtering Frequency Table by Dictionary/-ies}
\usage{
filterByDictionary(ft, fields = base_dictionary[fieldnames],
  fieldnames = c("Liebe"))
}
\arguments{
\item{ft}{A matrix as produced by \code{frequencytable()}.}

\item{fields}{A list of lists that contains the actual field names. 
By default, we load the base_dictionary (as in \code{dictionaryStatistics()}).}

\item{fieldnames}{A list of names for the dictionaries.}
}
\description{
This function can be used to filter a matrix as produced by 
\code{frequencytable()} by the words in the given dictionary(/-ies).
}
\examples{
data(rksp.0)
filtered <- filterByDictionary(frequencytable(rksp.0$mtext, byFigure = TRUE), fieldnames=c("Krieg", "Familie"))
}
