% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryStatistics.R
\name{filterFtByDict}
\alias{filterFtByDict}
\title{Filtering Frequency Table by Dictionary/-ies}
\usage{
filterFtByDict(ft, fields = loadFields(fieldnames, baseurl),
  fieldnames = c("Liebe"),
  baseurl = "https://raw.githubusercontent.com/quadrama/metadata/master/fields/")
}
\arguments{
\item{ft}{A matrix as produced by \code{frequencytable()}.}

\item{fields}{A list of lists that contains the actual field names. 
By default, we try to load the dictionaries using \code{fieldnames} and 
\code{baseurl} (as in \code{dictionaryStatistics()}).}

\item{fieldnames}{A list of names for the dictionaries.}

\item{baseurl}{The base path delivering the dictionaries.}
}
\description{
This function can be used to filter a matrix as produced by 
\code{frequencytable()} by the words in the given dictionary(/-ies).
}
\examples{
data(rksp.0)
filteredFt <- filterFtByDict(frequencytable(rksp.0$mtext, byFigure = TRUE), fieldnames=c("Krieg", "Familie"))
}
