% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{format.QDHasCharacter}
\alias{format.QDHasCharacter}
\title{Format Character Names}
\usage{
\method{format}{QDHasCharacter}(object, drama,
  FUN = stringr::str_to_title)
}
\arguments{
\item{object}{The object in which we want to transform names}

\item{drama}{The QDDrama object with all the information}

\item{FUN}{A function applied to the strings}
}
\description{
This function is applicable on all tables with a character table 
(that are in the class QDHasCharacter). It can be used to reformated the character 
names
}
\examples{
data(rksp.0)
ustat <- utteranceStatistics(rksp.0)
ustat <- format(ustat, rksp.0)
}
