% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencytable.R
\name{frequencytable2}
\alias{frequencytable2}
\title{Extract bigrams instead of words (currently not taking utterance boundaries into account)}
\usage{
frequencytable2(t, accepted.pos = postags$de$words, names = FALSE,
  cols = c("Token.surface", "Token.surface"), by.figure = FALSE)
}
\description{
Extract bigrams instead of words (currently not taking utterance boundaries into account)
}
