% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{generate_word_cloud}
\alias{generate_word_cloud}
\title{Generates a word cloud based on a frequency table}
\usage{
generate_word_cloud(freq_table, min.freq = 10)
}
\arguments{
\item{freq_table}{A single frequency table}

\item{min.freq}{The minimal frequency a token should have to be in the word cloud}
}
\description{
Generates a word cloud based on a frequency table
}
\examples{
t <- load_text("rksp.0", tokens=TRUE)
freq_table <- get_frequency_table(t)
generate_word_cloud(freq_table[1,1][[1]])

}

