% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{get_frequency_table}
\alias{get_frequency_table}
\title{Returns frequency tables for each figure in all dramas in t}
\usage{
get_frequency_table(t, relative = FALSE, accepted.pos = c())
}
\arguments{
\item{t}{A data frame representing (one or more) text}

\item{relative}{Whether to calculate relative frequencies}

\item{accepted.pos}{A vector containing pos tags that should be included.
If the vector is empty, includes everything.}
}
\description{
Returns frequency tables for each figure in all dramas in t
}
\examples{
t <- load_text("rksp.0", tokens=TRUE)
freq_table <- get_frequency_table(t)
}

