% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installData.R
\name{installCollectionData}
\alias{installCollectionData}
\title{Download and install collection data}
\usage{
installCollectionData(dataDirectory = getOption("qd.datadir"),
  branchOrCommit = "master", repository = "metadata",
  baseUrl = "https://github.com/quadrama/")
}
\arguments{
\item{dataDirectory}{The data directory in which collection and data files are stored}

\item{branchOrCommit}{The git branch, commit id, or tag that we want to download}

\item{repository}{The repository}

\item{baseUrl}{The github user (or group)}
}
\value{
\code{NULL}
}
\description{
Function to download collection data (grouped texts) from github. 
Overwrites (!) the current collections.
}
