% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installData.R
\name{installData}
\alias{installData}
\title{Download preprocessed drama data}
\usage{
installData(dataSource = "tg", dataDirectory = getOption("qd.datadir"),
  downloadSource = "ims", removeZipFile = TRUE)
}
\arguments{
\item{dataSource}{Currently, only "tg" (textgrid) is supported}

\item{dataDirectory}{The directory in which the data is to be stored}

\item{downloadSource}{The server from which to download}

\item{removeZipFile}{If true (the default), the downloaded zip file is removed after unpacking}
}
\description{
This function downloads pre-processed dramatic texts via http and stores them locally in your data directory
}
