% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{isolateFigureSpeech}
\alias{isolateFigureSpeech}
\title{Isolates Figure Speech}
\usage{
isolateFigureSpeech(drama, segment = c("Drama", "Act", "Scene"),
  min_token_count = 0, count_punctuation = TRUE,
  write_to_files = TRUE, dir = getOption("qd.datadir"))
}
\arguments{
\item{segment}{"Drama", "Act", or "Scene". Determines on what segment-level the speech is isolated.}

\item{min_token_count}{The minimal token count for a speech to be considered (default = 0)}

\item{count_punctuation}{Whether to include punctuation in min_token_count (default = TRUE)}

\item{write_to_files}{Whether to write each isolated speech into a new text file (default = TRUE)}

\item{dir}{The directory into which the files will be written (default = data directory)}

\item{t}{A text (or multiple texts, as data.frame or data.table)}
}
\description{
\code{isolateFigureSpeech()} isolates the speeches
of individual figures and optionally saves them in seperate text files.
}
\examples{
\dontrun{
data(rksp.0)
isolateFigureSpeech(rksp.0, segment="Scene")
}
}
