% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{limit.figures.by.rank}
\alias{limit.figures.by.rank}
\title{This method removes the spoken tokens of all but the most frequent n figures}
\usage{
limit.figures.by.rank(t, maxRank = 10)
}
\arguments{
\item{t}{The text, a data frame listing each token for each figure}

\item{maxRank}{Up to maxRank figures remain in the data set}
}
\description{
This method removes the spoken tokens of all but the most frequent n figures
}
