% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{limit.figures.by.tokens}
\alias{limit.figures.by.tokens}
\title{This method removes the spoken tokens by all figures that speak infrequently.}
\usage{
limit.figures.by.tokens(t, minTokens = 100)
}
\arguments{
\item{t}{The text, a data frame listing each token for each figure}

\item{minTokens}{The minimal amount of tokens a figure has to speak}
}
\description{
This method removes the spoken tokens by all figures that speak infrequently.
}
\examples{
data(rksp.0)
t <- limit.figures.by.tokens(rksp.0.text)
}
