% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{limitFiguresByTokens}
\alias{limitFiguresByTokens}
\title{This method removes the spoken tokens by all figures that speak infrequently.}
\usage{
limitFiguresByTokens(t, minTokens = 100, other = FALSE)
}
\arguments{
\item{t}{The text, a data frame listing each token for each figure}

\item{minTokens}{The minimal amount of tokens a figure has to speak}

\item{other}{Whether to summarize filtered figures as 'OTHER' instead of removing 
their speech. If it's of type character, it will be used as a name}
}
\description{
This method removes the spoken tokens by all figures that speak infrequently.
}
\keyword{internal}
