% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.annotations}
\alias{load.annotations}
\title{Helper method to load covered annotations.}
\usage{
load.annotations(ids, type = "de.unistuttgart.ims.drama.api.Utterance",
  coveredType = "de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token",
  url = "http://localhost:8080/drama.web")
}
\arguments{
\item{ids}{A vector or list of drama ids}

\item{type}{The annotation type to load}

\item{coveredType}{The annotation type of covered annotations we want to load}

\item{url}{The base url the web service can be reached with}
}
\description{
Helper method to load covered annotations.
}
\examples{
\dontrun{
load.annotations(c("rksp.0"))
}
}
