% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.sets}
\alias{load.sets}
\title{A function to get a list of all collections and the number of plays in that collection}
\usage{
load.sets()
}
\description{
A function to get a list of all collections and the number of plays in that collection
}
