% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.text}
\alias{load.text}
\title{Loads a CSV-formatted text from the server,
assuming the main server url has been set correctly.}
\usage{
load.text(ids, tokens = FALSE, url = "http://localhost:8080/drama.web")
}
\arguments{
\item{ids}{A vector containing drama ids to be downloaded}

\item{tokens}{If set to true, the table also contains each token in an utterance}

\item{url}{The url we load the text from}
}
\description{
Loads a CSV-formatted text from the server,
assuming the main server url has been set correctly.
}
