% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.text2}
\alias{load.text2}
\title{Similar to load.text(), but the table also includes scene and act markings.}
\usage{
load.text2(ids)
}
\arguments{
\item{ids}{The ids for which we want to get the text}
}
\description{
Similar to load.text(), but the table also includes scene and act markings.
}
\examples{
\dontrun{
mtext <- load.text2(c("tg:rksp.0"))
}
}
