% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_text}
\alias{load_text}
\title{Loads a CSV-formatted text from the server,
assuming the main server url has been set correctly.}
\usage{
load_text(..., tokens = FALSE)
}
\description{
Loads a CSV-formatted text from the server,
assuming the main server url has been set correctly.
}

