% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadAllInstalledIds}
\alias{loadAllInstalledIds}
\title{Installed texts}
\usage{
loadAllInstalledIds(asDataFrame = FALSE,
  dataDirectory = getOption("qd.datadir"))
}
\arguments{
\item{asDataFrame}{Logical value. Controls whether 
the return value is a list (with colon-joined ids) 
or a data.frame with two columns (corpus, drama)}

\item{dataDirectory}{The directory in which precompiled 
drama data is installed}
}
\description{
Returns a list of all ids that are installed
}
