% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadAnnotations}
\alias{loadAnnotations}
\title{Load annotations
Helper method to load covered annotations.}
\usage{
loadAnnotations(ids, type = "de.unistuttgart.ims.drama.api.Utterance",
  coveredType = "de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token")
}
\arguments{
\item{ids}{A vector or list of drama ids}

\item{type}{The annotation type to load}

\item{coveredType}{The annotation type of covered annotations we want to load}
}
\examples{
\dontrun{
loadAnnotations(c("rksp.0"))
}
}
