% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadCharacters}
\alias{loadCharacters}
\title{Character data loading}
\usage{
loadCharacters(
  ids,
  defaultCollection = "tg",
  dataDirectory = getOption("qd.datadir")
)
}
\arguments{
\item{ids}{a list or vector of ids}

\item{defaultCollection}{the default collection}

\item{dataDirectory}{the data directory}
}
\value{
A data.frame extracted from the CSV file about characters
}
\description{
Loads a table of characters and meta data
}
