% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTEI.R
\name{loadDramaTEI}
\alias{loadDramaTEI}
\title{Load drama}
\usage{
loadDramaTEI(filename)
}
\arguments{
\item{filename}{The filename of the drama to load (or a list thereof).}
}
\value{
The function returns an object of class \code{QDDrama}.
}
\description{
This function parses and loads one or more dramas in raw TEI format.
}
\examples{
\donttest{
tei_example_file <- "example_tei.xml"
d <- loadDramaTEI(tei_example_file)
}

}
