% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadMeta}
\alias{loadMeta}
\title{Load meta data}
\usage{
loadMeta(ids)
}
\arguments{
\item{ids}{A vector or list of drama ids}
}
\value{
a data frame
}
\description{
helper method to load meta data about dramatic texts (E.g., author, year). 
Does not load the texts, so it's much faster.
}
