% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadNumbers}
\alias{loadNumbers}
\title{This function loads the number of annotations of different types for several 
dramas at once.}
\usage{
loadNumbers(ids = c(), types = c("de.unistuttgart.ims.drama.api.Act",
  "de.unistuttgart.ims.drama.api.Scene",
  "de.unistuttgart.ims.drama.api.Utterance",
  "de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token",
  "de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence",
  "de.unistuttgart.ims.drama.api.DramatisPersonae"), debug = FALSE)
}
\arguments{
\item{ids}{A vector containing drama ids}

\item{types}{A character vector containing the annotation types we want to count}

\item{debug}{Logical value, whether to print out debug info}
}
\description{
This function loads the number of annotations of different types for several 
dramas at once.
}
\examples{
\dontrun{
loadNumbers(c("rksp.0", "vndf.0"))
}
}
