% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadSet}
\alias{loadSet}
\title{Function to load a set from QuaDramA web service.
Can optionally set the set name as a genre in the returned table}
\usage{
loadSet(setName, addGenreColumn = FALSE)
}
\arguments{
\item{setName}{The name of the set to retrieve}

\item{addGenreColumn}{Whether to set the Genre-column in the returned table to the set name}
}
\description{
Function to load a set from QuaDramA web service.
Can optionally set the set name as a genre in the returned table
}
