% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{mapPrefix}
\alias{mapPrefix}
\title{Replace corpus prefix}
\usage{
mapPrefix(idList, map)
}
\arguments{
\item{idList}{The list of ids in which we want to replace.}

\item{map}{A list containing the old prefix as name and the new one as
values.}
}
\description{
This function can be used to replace corpus prefixes.
If a list of play ids contains textgrid prefixes, for instance, this 
function can be used to map them onto GerDraCor prefixes. Please note
that the function does \emph{not} check whether the play actually exists 
in the corpus.
}
\examples{

# returns c("corpus2:play1", "corpus2:play2")
mapPrefix(c("corpus1:play1", "corpus1:play2"), list(corpus1="corpus2"))

}
