% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{numberOfPlays}
\alias{numberOfPlays}
\title{Number of plays}
\usage{
numberOfPlays(x)
}
\arguments{
\item{x}{The QDDrama object}
}
\value{
An integer. The number of plays contained in the QDDrama object.
}
\description{
The function \code{numberOfPlays()} determines how many
different plays are contained in a single QDDrama object.
}
\examples{
# returns 1
numberOfPlays(rksp.0)

# returns 2
numberOfPlays(combine(rksp.0, rjmw.0))
}
