% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.QDHamming}
\alias{plot.QDHamming}
\title{Personnel Exchange}
\usage{
\method{plot}{QDHamming}(x, drama = NULL, xlab = "Scene",
  ylab = "Exchange after Scene", ...)
}
\arguments{
\item{x}{A numeric vector generated from the function}

\item{drama}{Optional QDDrama object. If present, act boundaries
and correct scene labels are included in the plot.}

\item{xlab}{A character vector that is used as x axis label. Defaults to
"Scene".}

\item{ylab}{A character vector that is used as y axis label. Defaults to
"Exchange".}

\item{...}{Parameters passed to \code{plot.default()}.}
}
\value{
See \code{plot.default()}.
}
\description{
Uses the default scatterplot function to plot the personnel 
exchange in each scene.
}
\examples{
data(rksp.0)
h <- hamming(rksp.0)
plot(h, drama=rksp.0)
}
\seealso{
plot.default
}
