% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.QDUtteranceStatistics}
\alias{plot.QDUtteranceStatistics}
\title{Utterance positions}
\usage{
\method{plot}{QDUtteranceStatistics}(x, drama = NULL,
  colors = qd.colors, xlab = "Time", ...)
}
\arguments{
\item{x}{A table generated from the function}

\item{drama}{Optional QDDrama object. If present, segment boundaries
are extracted from it and included in the plot.}

\item{colors}{The colors to be used}

\item{xlab}{A character vector that is used as x axis label. Defaults to
"Time".}

\item{...}{Parameters passed to stripchart().}
}
\value{
See \code{stripchart()}.
}
\description{
Uses the function \code{stripchart} to  plot each utterance at their position,
in a line representing the character. The dot is marked in the middle of each utterance.
Might look weird if very long utterances are present.
}
\seealso{
stripchart
}
