% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotUtterancePositions}
\alias{plotUtterancePositions}
\title{Utterance positions}
\usage{
plotUtterancePositions(utteranceStatistics, segmentedText = NULL,
  colors = qd.colors, plotFrame = FALSE, ...)
}
\arguments{
\item{utteranceStatistics}{A table generated from the function}

\item{segmentedText}{If supplied, act boundaries will be marked}

\item{colors}{The colors to be used}

\item{plotFrame}{Whether to draw a frame around the plot}

\item{...}{Parameters passed to stripchart().}
}
\description{
Uses `stripchart` to  plot each utterance at their position,
in a line representing the figure. The dot is marked in the middle of each utterance.
Might look weird if very long utterances are present.
}
