% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ling.R
\docType{data}
\name{postags}
\alias{postags}
\title{Provides lists of groups of pos tags for various word classes.}
\format{An object of class \code{list} of length 1.}
\usage{
postags
}
\description{
Provides lists of groups of pos tags for various word classes.
}
\keyword{datasets}
