% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\docType{data}
\name{qd.colors}
\alias{qd.colors}
\title{QuaDramA colors}
\format{
An object of class \code{character} of length 10.
}
\usage{
qd.colors
}
\description{
color scheme to be used for QuaDramA plots
Taken from http://google.github.io/palette.js/, tol-rainbow, 10 colors
}
\keyword{datasets}
