% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{rank.figures.by.appearance}
\alias{rank.figures.by.appearance}
\title{Given a dramatic text and a table of figures, ranks the figures by 
their first appearance. The lower the rank number, the earlier the figure appears.
"appears": speaks for the first time.}
\usage{
rank.figures.by.appearance(figures, text,
  columnTitle = "Rank (1st appearance)")
}
\arguments{
\item{figures}{A data frame containing the figures}

\item{text}{A text data frame}
}
\description{
Given a dramatic text and a table of figures, ranks the figures by 
their first appearance. The lower the rank number, the earlier the figure appears.
"appears": speaks for the first time.
}
\examples{
data(rksp.0.text, rksp.0.figures)
rank.figures.by.appearance(rksp.0.figures, rksp.0.text)
}
