% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{rankFiguresByAppearance}
\alias{rankFiguresByAppearance}
\title{Rank figures by their 1st appearance}
\usage{
rankFiguresByAppearance(figures, text, columnTitle = "Rank (1st appearance)")
}
\arguments{
\item{figures}{A data frame containing the figures}

\item{text}{A text data frame}

\item{columnTitle}{The title for the rank column}
}
\description{
Given a dramatic text and a table of figures, ranks the figures by 
their first appearance. The lower the rank number, the earlier the figure appears.
"appears": speaks for the first time.
}
