% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{rankFiguresByDramatisPersonae}
\alias{rankFiguresByDramatisPersonae}
\title{Adds a column to the figures data frame, containing the rank in the dramatis personae.}
\usage{
rankFiguresByDramatisPersonae(figures,
  columnTitle = "Rank (dramatis personae)")
}
\arguments{
\item{figures}{The figures to rank}

\item{columnTitle}{The title for the rank column}
}
\description{
Adds a column to the figures data frame, containing the rank in the dramatis personae.
}
