% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{report}
\alias{report}
\title{Report}
\usage{
report(id = "tg:rksp.0", of = file.path(getwd(), paste0(unlist(strsplit(id,
  ":", fixed = TRUE))[2], "html")), colors = qd.colors)
}
\arguments{
\item{id}{The id of the text}

\item{of}{The output file}

\item{colors}{A list of colors to be used for plots}
}
\description{
generates a report for a specific dramatic text
}
