% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rksp.0.figures}
\alias{rksp.0.figures}
\title{Figures declared in the dramatis personae from Emilia Galotti.}
\format{An object of class \code{data.frame} with 10 rows and 6 columns.}
\usage{
rksp.0.figures
}
\description{
Figures declared in the dramatis personae from Emilia Galotti.
}
\keyword{datasets}

