% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rksp.0.mtext}
\alias{rksp.0.mtext}
\title{The data set exported from Emilia Galotti}
\format{A data frame with 25365 rows and 19 columns:
\describe{
 \item{drama}{The drama id}
 \item{length.Act}{The number of acts}
 \item{begin.Act}{The begin position of the current act}
 \item{end.Act}{The end position of the current act}
 \item{Number.Act}{The number of the current act}
 \item{Regular.Act}{A boolean indicating wether this act was created for technical reasons}
 \item{length.Scene}{The number of scenes}
 \item{begin.Scene}{The begin position of the current scene}
 \item{end.Scene}{The end position of the current scene}
 \item{Number.Scene}{The number of the current scene}
 \item{Regular.Scene}{A boolean indicating wether this scene was created for technical reasons}
 \item{length}{The length of the drama (in tokens)}
 \item{begin}{The beginning of an utterance (character position)}
 \item{end}{The end of an utterance (character position)}
 \item{Speaker.figure_surface}{The surface description of a figure}
 \item{Speaker.figure_id}{}
 \item{Token.surface}{}
 \item{Token.pos}{}
 \item{Token.lemma}{}
}}
\usage{
rksp.0.mtext
}
\description{
The data set exported from Emilia Galotti
}
\keyword{datasets}
