% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rksp.0}
\alias{rksp.0}
\title{The data set exported from Emilia Galotti}
\format{A list containing data.frames and data.table.
The key "mtext" contains the following data.table:}
\usage{
rksp.0
}
\description{
The data set exported from Emilia Galotti
}
\keyword{datasets}
