% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveCollection}
\alias{saveCollection}
\title{Save or Extend a Collection}
\usage{
saveCollection(t, name = ifelse(typeof(t) == "character", paste(t,
  collapse = "_"), paste(unique(t$drama), collapse = "_")),
  dir = getOption("qd.collectionDirectory"), append = TRUE)
}
\arguments{
\item{t}{A text (or multiple texts, as data.frame or data.table), or a character 
vector containing the drama IDs to be saved}

\item{name}{The name of the collection file (default: concatenated drama IDs)}

\item{dir}{The directory into which the collection file will be written (uses the 
collection directory as initialized by \code{setup()} by default)}

\item{append}{Whether to extend the collection file if it already exists. 
If FALSE, the file will be overwritten (default: TRUE).}
}
\description{
\code{saveCollection()} can be used to create new collections 
or add dramas to existing collection files.
}
\examples{
\dontrun{
data(rksp.0, rjmw.0)
saveCollection(rbind(rksp.0$mtext, rjmw.0$mtext))
saveCollection(c("rksp.0", "rjmw.0"), append=FALSE) # produces identical file
saveCollection(c("a", "b"), name="rksp.0_rjmw.0") # adds "a" and "b" to the file
}
}
