% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveFigureSpeech}
\alias{saveFigureSpeech}
\title{Save Isolated Figure Speech}
\usage{
saveFigureSpeech(t, segment = c("Drama", "Act", "Scene"),
  min_token_count = 0, count_punctuation = TRUE, dir = getwd())
}
\arguments{
\item{t}{A text (or multiple texts, as data.frame or data.table)}

\item{segment}{"Drama", "Act", or "Scene". Determines on what segment-level the speech is isolated.}

\item{min_token_count}{The minimal token count for a speech to be saved in a file (default = 0)}

\item{count_punctuation}{Whether to include punctuation in min_token_count}

\item{dir}{The directory into which the files will be written (default: current working directory)}
}
\description{
\code{saveFigureSpeech()} isolates the speeches
of individual figures and saves them in seperate text files.
}
\examples{
\dontrun{
data(rksp.0)
saveFigureSpeech(rksp.0$mtext, segment="Scene")
}
}
