% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installData.R, R/load.R
\name{installCollectionData}
\alias{installCollectionData}
\alias{installData}
\alias{setup}
\alias{setDataDirectory}
\alias{setCollectionDirectory}
\title{Download and install collection data}
\usage{
installCollectionData(dataDirectory = getOption("qd.datadir"),
  branchOrCommit = "master", repository = "metadata",
  baseUrl = "https://github.com/quadrama/")

installData(dataSource = "tg", dataDirectory = getOption("qd.datadir"),
  downloadSource = "ims", removeZipFile = TRUE,
  baseUrl = "https://github.com/quadrama", remoteUrl = paste0(baseUrl,
  "/data_", dataSource, ".git"))

setup(dataDirectory = file.path(path.expand("~"), "QuaDramA", "Data2"),
  collectionDirectory = file.path(dataDirectory, "collections"))

setDataDirectory(dataDirectory = file.path(path.expand("~"), "QuaDramA",
  "Data2"))


  setCollectionDirectory(collectionDirectory = file.path(getOption("qd.datadir"),
  "collections"))
}
\arguments{
\item{dataDirectory}{The data directory in which collection and data files are stored}

\item{branchOrCommit}{The git branch, commit id, or tag that we want to download}

\item{repository}{The repository}

\item{baseUrl}{The github user (or group)}

\item{dataSource}{Currently, only "tg" (textgrid) is supported}

\item{downloadSource}{No longer used.}

\item{removeZipFile}{No longer used.}

\item{remoteUrl}{The URL of the remote repository.}

\item{collectionDirectory}{A path to the directory in which collections are stored. 
By default, the directory is called "collection" below the data directory.}

\item{dataDirectory}{The directory in which the data is to be stored}

\item{baseUrl}{The remote repository owner (e.g., https://github.com/quadrama)}

\item{dataDirectory}{A path to the directory in which data and metadata are located. 
"~/QuaDramA/Data2" by default.}
}
\description{
Function to download collection data (grouped texts) from github. 
Overwrites (!) the current collections.

This function downloads pre-processed dramatic texts via http and stores them locally in your data directory
}
