% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{split.QDDrama}
\alias{split.QDDrama}
\title{Split multiple plays}
\usage{
\method{split}{QDDrama}(x, ...)
}
\arguments{
\item{x}{The object of class \code{QDDrama} (consisting of multiple dramas). 
For \code{split()} it should consist of multiple plays. For \code{combine()} it 
can but doesn't have to.}

\item{...}{All other arguments are ignored.}
}
\value{
Returns a list of individual QDDrama objects, each containing one text.
}
\description{
The function \code{split(x)} expects an object of type \code{QDDrama} and can 
be used to split a \code{QDDrama} object that consists of multiple dramas 
into a list thereof. It is the counterpart to \code{combine(x, y)}.
}
\examples{
data(rksp.0)
data(rjmw.0)
d <- combine(rjmw.0, rksp.0)
dlist <- split(d)
}
