% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{tail}
\alias{tail}
\title{Extract section}
\usage{
tail(input, by = c("Act", "Scene"), op = "==", n = 1)
}
\arguments{
\item{input}{Segmented text (can be multiple texts)}

\item{by}{Act or Scene, or matching substring}

\item{op}{Whether to extract exactly one or more than one}
}
\examples{
data(rksp.0)
# Extract the second last scene
tail(rksp.0$mtext, by="Scene", op="==", n=2)
}
