% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utterance_statistics.R
\name{utterance_statistics}
\alias{utterance_statistics}
\title{This method calculates the length of each utterance, organised by figure and drama.}
\usage{
utterance_statistics(t, num.figures = 10, normalize.by.drama.length = TRUE)
}
\arguments{
\item{t}{The dramatic text(s)}

\item{num.figures}{The maximal number of figures per drama to include. Default: 10. Set to FALSE to include all figures.}
}
\description{
This method calculates the length of each utterance, organised by figure and drama.
}
\examples{
data(rksp.0)
num_figures <- 5
ustat <- utterance_statistics(rksp.0, num.figures = num_figures)
boxplot(ustat$utterance_length ~ ustat$figure,col=qd.colors[1:num_figures], las=2,frame=FALSE)
}

