% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{configuration}
\alias{configuration}
\title{Creates classic drama configuration matrix. Returns a list with 
the three components matrix, drama, and figure}
\usage{
configuration(mtext, by = "Act", onlyPresence = FALSE)
}
\arguments{
\item{mtext}{The text including Act and Scene markings}

\item{by}{A string, either "Act" or "Scene"}

\item{onlyPresence}{If TRUE, the resulting matrix only contains 
logical values for stage presence}
}
\description{
Creates classic drama configuration matrix. Returns a list with 
the three components matrix, drama, and figure
}
\examples{
data(rksp.0)
cfg <- configuration(rksp.0$mtext)

}
\seealso{
DramaAnalysis::load.text2()
}
