% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{configuration}
\alias{configuration}
\title{Creates classic drama configuration matrix}
\usage{
configuration(mtext, by = "Act", only.presence = FALSE)
}
\arguments{
\item{mtext}{The text including Act and Scene markings}

\item{by}{A string, either "Act" or "Scene"}
}
\description{
Creates classic drama configuration matrix
}
\examples{
data(rksp.0.mtext)
cfg <- configuration(rksp.0.mtext)

}
\seealso{
DramaAnalysis::load.text2()
}
