% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{configuration}
\alias{configuration}
\title{Creates classic drama configuration matrix. Returns a list with 
the three components matrix, drama, and figure}
\usage{
configuration(mtext, by = c("Act", "Scene"), onlyPresence = FALSE,
  useCharacterId = FALSE)
}
\arguments{
\item{mtext}{The text including Act and Scene markings}

\item{by}{A string, either "Act" or "Scene". Partial matching allowed.}

\item{onlyPresence}{If TRUE, the resulting matrix only contains 
logical values for stage presence}

\item{useCharacterId}{Logical. If true, characters are represented 
by their id instead of their surface form.}
}
\description{
Creates classic drama configuration matrix. Returns a list with 
the three components matrix, drama, and figure
}
\examples{
data(rksp.0)
cfg <- configuration(rksp.0$mtext)

}
