% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{rksp.0}
\alias{rjmw.0}
\title{Data sets}
\format{
A list containing data.frames and data.table.

An object of class \code{QDDrama} (inherits from \code{list}) of length 6.
}
\usage{
rksp.0

rjmw.0
}
\description{
\code{rksp.0} represents the data set exported from Lessings 
Emilia Galotti, \code{rjmw.0} is the one exported from Miss Sara Sampson 
(also written by Lessing). Please note that in both plays, special 
characters have been removed for technical reasons. The text is German, 
but all umlauts have been replaced by another character. This is only a 
restriction of the pre-packaged files.
}
\keyword{datasets}
