% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.statistics.R
\name{dictionary.statistics}
\alias{dictionary.statistics}
\title{This method retrieves word lists from the given URL (e.g., github) and
counts the number of occurrences of the words in the dictionaries}
\usage{
dictionary.statistics(t, fieldnames = c(), normalize.by.figure = FALSE,
  normalize.by.field = FALSE, names = FALSE, boost = 1,
  baseurl = "https://raw.githubusercontent.com/quadrama/metadata/master/fields/")
}
\arguments{
\item{t}{A text}

\item{fieldnames}{A list of names for the dictionaries. It is expected that files with that name can be found below the URL.}

\item{normalize.by.figure}{Whether to normalize by figure speech length}

\item{normalize.by.field}{Whether to normalize by dictionary size. You usually want this.}

\item{names}{Whether the resulting table contains figure ids or names}

\item{boost}{A scaling factor to generate nicer values}

\item{baseurl}{The url delivering the dictionairies}
}
\description{
This method retrieves word lists from the given URL (e.g., github) and
counts the number of occurrences of the words in the dictionaries
}
