% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryStatistics.R
\name{dictionaryStatistics}
\alias{dictionaryStatistics}
\title{This method retrieves word lists from the given URL (e.g., github) and
counts the number of occurrences of the words in the dictionaries}
\usage{
dictionaryStatistics(t, fieldnames = c(), normalizeByFigure = FALSE,
  normalizeByField = FALSE, names = FALSE, boost = 1,
  baseurl = "https://raw.githubusercontent.com/quadrama/metadata/master/fields/",
  column = "Token.surface")
}
\arguments{
\item{t}{A text}

\item{fieldnames}{A list of names for the dictionaries. It is expected that files with that name can be found below the URL.}

\item{normalizeByFigure}{Whether to normalize by figure speech length}

\item{normalizeByField}{Whether to normalize by dictionary size. You usually want this.}

\item{names}{Whether the resulting table contains figure ids or names}

\item{boost}{A scaling factor to generate nicer values}

\item{baseurl}{The base path delivering the dictionaries. Should end in a /, field names will be appended and fed into read.csv().}

\item{column}{The table column we apply the dictionary on. Should be either "Token.surface" or "Token.lemma".}
}
\description{
This method retrieves word lists from the given URL (e.g., github) and
counts the number of occurrences of the words in the dictionaries
}
