% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryStatistics.R
\name{dictionaryStatisticsSingle}
\alias{dictionaryStatisticsSingle}
\title{Calculates dictionary statistics for a single word field that is given as a character vector}
\usage{
dictionaryStatisticsSingle(t, wordfield = c(), names = FALSE,
  normalizeByFigure = FALSE, normalizeByField = FALSE,
  fieldNormalizer = length(wordfield), bylist = ifelse(names == TRUE,
  "corpus,drama,Speaker.figure_surface", "corpus,drama,Speaker.figure_id"),
  column = "Token.surface", ci = TRUE, colnames = c("corpus", "drama",
  "figure", "x"))
}
\arguments{
\item{t}{A single or multiple text(s)}

\item{wordfield}{A character vector containing the words or lemmas to be counted}

\item{names}{A logical values. Whether to use figure names or ids}

\item{normalizeByFigure}{A logical value. Whether to normalize by the amount of tokens a figure speaks}

\item{normalizeByField}{A logical value. Whether to normalize by the size of the word field}

\item{fieldNormalizer}{defaults to the length of the wordfield}

\item{bylist}{A list of columns, to be passed into the aggregate function. Can be used to control whether to count by figures or by dramas}

\item{column}{"Token.surface" or "Token.lemma"}

\item{ci}{Logical. Wether to ignore case. Defaults to TRUE}

\item{colnames}{The column names to be used}
}
\description{
Calculates dictionary statistics for a single word field that is given as a character vector
}
\examples{
data(rksp.0.text)
fstat <- dictionaryStatisticsSingle(rksp.0.text, wordfield=c("der"), names=TRUE)
}
