% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencytable.R
\name{frequencytable}
\alias{frequencytable}
\title{This function generates a table to be used by stylo}
\usage{
frequencytable(t, accepted.pos = postags$de$words, names = FALSE,
  column = "Token.surface", by.figure = FALSE)
}
\arguments{
\item{t}{The text table, potentially covering multiple texts}

\item{accepted.pos}{A list of accepted pos tags}

\item{names}{Whether to use figure names or ids}

\item{column}{The column name we should use (should be either Token.surface or Token.lemma)}

\item{by.figure}{Wether the count is by figure or by text}
}
\description{
This function generates a table to be used by stylo
}
\examples{
data(rksp.0)
st <- frequencytable(rksp.0)
}

