% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryStatistics.R, R/frequencytable.R
\name{filterByDictionary}
\alias{filterByDictionary}
\alias{frequencytable}
\title{Word frequencies}
\usage{
filterByDictionary(ft, fields = base_dictionary[fieldnames],
  fieldnames = c("Liebe"))

frequencytable(drama, acceptedPOS = postags$de$words,
  column = "Token.lemma", byFigure = FALSE, sep = "|",
  normalize = FALSE, sortResult = FALSE, segment = c("Drama", "Act",
  "Scene"))
}
\arguments{
\item{ft}{A matrix as produced by \code{frequencytable()}.}

\item{fields}{A list of lists that contains the actual field names. 
By default, we load the base_dictionary (as in \code{dictionaryStatistics()}).}

\item{fieldnames}{A list of names for the dictionaries.}

\item{drama}{A \code{QDDrama}. May be covering multiple texts}

\item{acceptedPOS}{A list of accepted pos tags}

\item{column}{The column name we should use (should be either Token.surface or Token.lemma)}

\item{byFigure}{Wether the count is by figure or by text}

\item{sep}{The separation character that goes between drama name and figure (if applicable)}

\item{normalize}{Whether to normalize values or not}

\item{sortResult}{Logical. If true, the columns with the highest sum are ordered left (i.e., frequent words are visible first)}

\item{segment}{Whether the count is by drama (default), act or scene}
}
\description{
The function \code{filterByDictionary()} can be used to filter a matrix as produced by 
\code{frequencytable()} by the words in the given dictionary(/-ies).

The function \code{frequencytable()} generates a matrix of word frequencies 
by drama, act or scene and/or by figure.
}
\examples{
data(rksp.0)
filtered <- filterByDictionary(frequencytable(rksp.0, 
                                              byFigure = TRUE), 
                                              fieldnames=c("Krieg", "Familie"))

data(rksp.0)
st <- frequencytable(rksp.0)
\dontrun{
stylo(gui=F, frequencies = st)
}
}
