% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencytable.R
\name{frequencytable}
\alias{frequencytable}
\title{This function generates a matrix of word frequencies by figure or drama.}
\usage{
frequencytable(t, accepted.pos = postags$de$words, names = FALSE,
  column = "Token.surface", by.figure = FALSE, sep = "|")
}
\arguments{
\item{t}{The text table, potentially covering multiple texts}

\item{accepted.pos}{A list of accepted pos tags}

\item{names}{Whether to use figure names or ids}

\item{column}{The column name we should use (should be either Token.surface or Token.lemma)}

\item{by.figure}{Wether the count is by figure or by text}

\item{sep}{The separation character that goes between drama name and figure (if applicable)}
}
\description{
This function generates a matrix of word frequencies by figure or drama.
}
\examples{
data(rksp.0.text)
st <- frequencytable(rksp.0.text)
\dontrun{
t <- load.text(read.csv("http://localhost:8080/drama.web/dramas", header=FALSE)[,], tokens=T)
tl <- limit.figures.by.tokens(t, minTokens=1000)
stylo_table <- frequencytable(tl, names=TRUE, by.figure=TRUE)
stylo(gui=F, frequencies = stylo_table, network=T, write.png.file=T, analysis.type="BCT")
}
}
