% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencytable.R
\name{frequencytable2}
\alias{frequencytable2}
\title{Extract bigrams instead of words (currently not taking utterance boundaries into account)}
\usage{
frequencytable2(t, acceptedPOS = postags$de$words, names = FALSE,
  column = c("Token.surface", "Token.surface"), byCharacter = FALSE,
  segment = c("Drama", "Act", "Scene"))
}
\arguments{
\item{t}{The text}

\item{acceptedPOS}{A list of accepted pos tags}

\item{names}{Whether to use character names or ids}

\item{column}{The column names we should use (should be either Token.surface or Token.lemma)}

\item{byCharacter}{Wether the count is by character or by text}

\item{segment}{Whether the count is by drama (default), act or scene}
}
\value{
Matrix of bigram frequencies in the format bigrams X segments
}
\description{
Extract bigrams instead of words (currently not taking utterance boundaries into account)
}
\keyword{internal}
