% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{installData}
\alias{installData}
\title{Download preprocessed drama data}
\usage{
installData(dataSource = "tg", dataDirectory = getOption("qd.datadir"),
  downloadSource = "ims")
}
\arguments{
\item{dataSource}{Currently, only "tg" (textgrid) is supported}

\item{dataDirectory}{The directory in which the data is to be stored}

\item{downloadSource}{The server from which to download}
}
\description{
This function downloads pre-processed dramatic texts via http and stores them locally in your data directory
}
