% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness.R
\name{keyness}
\alias{keyness}
\title{Log-Likelihood}
\usage{
keyness(ft, row = 1, epsilon = 1e-100, siglevel = 0.05,
  method = "loglikelihood", minimalFrequency = 10)
}
\arguments{
\item{ft}{The frequency table}

\item{row}{The row number we want to compare to the others}

\item{epsilon}{null values are replaced by this value, in order to avoid division by zero}

\item{siglevel}{Return only the keywords above the significance level. Set to 1 to get all words}
}
\value{
A list of keywords, sorted by their log-likelihood value, calculated according to http://ucrel.lancs.ac.uk/llwizard.html
}
\description{
Given a frequency table (with texts as rows and words as columns),
this function calculates the log-likelihood of one row against all other rows. 
The return value is a matrix containing log-likelihood scores for each word
}
