% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.text2}
\alias{load.text2}
\title{Loads a CSV-formatted text from the given server url.}
\usage{
load.text2(ids, url = "http://localhost:8080/drama.web")
}
\arguments{
\item{ids}{A list or vector of ids}

\item{url}{the URL to reach the server.}
}
\description{
Loads a CSV-formatted text from the given server url.
}
