% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.text2}
\alias{load.text2}
\title{Loads a CSV-formatted text from the given server url.
This function is an incredible bad idea.}
\usage{
load.text2(ids)
}
\description{
Loads a CSV-formatted text from the given server url.
This function is an incredible bad idea.
}
