% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadAnnotations}
\alias{loadAnnotations}
\title{Load annotations}
\usage{
loadAnnotations(ids, type = atypes$Utterance, coveredType = atypes$Token,
  defaultCollection = "tg", columnTypes = NULL)
}
\arguments{
\item{ids}{A vector or list of drama ids}

\item{type}{The annotation type to load}

\item{coveredType}{The annotation type of covered annotations we want to load}

\item{defaultCollection}{The collection prefix is added if no prefix is found}

\item{columnTypes}{Can be used to specify column types, which are passed to readr::read.csv.}
}
\description{
Helper method to load covered annotations. Returns a data.table.
}
\examples{
\dontrun{
loadAnnotations(c("tg:rksp.0"))
}
}
