% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R, R/text.R
\name{loadDrama}
\alias{loadDrama}
\alias{combine}
\title{Load drama}
\usage{
loadDrama(ids, defaultCollection = "qd")

combine(x, y)
}
\arguments{
\item{ids}{A vector of ids.}

\item{defaultCollection}{If the ids do not have a collection prefix, the 
defaultCollection prefix is applied.}

\item{x}{A \code{QDDrama}}

\item{y}{A \code{QDDrama}}
}
\value{
The function returns a \code{QDDrama} object. This is essentially a 
list of \code{data.table}s, covering the different aspects (utterances, segments, 
characters, ...). If multiple ids have been supplied as arguments, the tables 
contain the information of multiple plays.
}
\description{
This function loads one or more of the installed plays and 
returns them as a \code{QDDrama} object.

The function \code{combine(x, y)} can be used to merge 
multiple objects of the type \code{QDDrama} into one.
}
\examples{
\dontrun{
# both are equivalent
d <- loadDrama(c("test:rksp.0", "test:rjmw.0"))
d <- loadDrama(c("rksp.0", "rjmw.0"), defaultCollection = "test")
}

data(rksp.0)
data(rjmw.0)
d <- combine(rjmw.0, rksp.0)
}
