% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTEI.R
\name{loadDramaTEI}
\alias{loadDramaTEI}
\title{Load drama}
\usage{
loadDramaTEI(filename, dataDirectory = paste0(getOption("qd.datadir"),
  "/tei"))
}
\arguments{
\item{filename}{The filename of the drama to load (or a list thereof).}

\item{dataDirectory}{The directory that holds the file(s).}
}
\description{
This function parses and loads one or more dramas in raw TEI format.
}
