% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadMeta}
\alias{loadMeta}
\title{Load meta data}
\usage{
loadMeta(ids, type = NULL)
}
\arguments{
\item{ids}{A vector or list of drama ids}

\item{type}{The annotation type to load. No longer used.}
}
\description{
helper method to load meta data about dramatic texts (E.g., author, year)
}
