% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadNumbers}
\alias{loadNumbers}
\title{This function loads the number of annotations of different types for several 
dramas at once.}
\usage{
loadNumbers(ids = c(), types = c(atypes$Act, atypes$Scene, atypes$Utterance,
  atypes$Token, atypes$Sentence, atypes$DramatisPersonae), debug = FALSE)
}
\arguments{
\item{ids}{A vector containing drama ids}

\item{types}{A character vector containing the annotation types we want to count}

\item{debug}{Logical value, whether to print out debug info}
}
\description{
This function loads the number of annotations of different types for several 
dramas at once.
}
\examples{
\dontrun{
loadNumbers(c("rksp.0", "vndf.0"))
}
}
