% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadSegmentedText}
\alias{loadSegmentedText}
\title{Similar to load.text(), but the table also includes scene and act markings.}
\usage{
loadSegmentedText(ids)
}
\arguments{
\item{ids}{The ids for which we want to get the text}
}
\description{
Similar to load.text(), but the table also includes scene and act markings.
}
\examples{
\dontrun{
mtext <- loadSegmentedText("tg:rksp.0")
}
}
