% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadSegmentedText}
\alias{loadSegmentedText}
\title{Text Loading}
\usage{
loadSegmentedText(ids, defaultCollection = "tg")
}
\arguments{
\item{ids}{The ids for which we want to get the text}

\item{defaultCollection}{The collection prefix is added if no prefix is found}
}
\description{
Loads a text with its segmentation
}
\examples{
\dontrun{
installData("test")
mtext <- loadSegmentedText("test:rksp.0")
}
}
