% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{loadSet}
\alias{loadSet}
\alias{loadSets}
\title{Load Collections}
\usage{
loadSet(setName, addGenreColumn = FALSE)

loadSets()
}
\arguments{
\item{setName}{A character vector. The name of the set(s) to retrieve.}

\item{addGenreColumn}{Logical. Whether to set the Genre-column in 
the returned table to the set name. If set to FALSE (default), a vector
is returned. In this case, association to collections is not returned.
Otherwise, it's a data.frame.}
}
\description{
Function to load a set from collection files
Can optionally set the set name as a genre in the returned table. 
\code{loadSets()} returns table of all defined collections (and the
number of plays in each).
}
