% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadText}
\alias{loadText}
\title{Loads a CSV-formatted text from the server,
assuming the main server url has been set correctly.}
\usage{
loadText(ids, includeTokens = FALSE, defaultCollection = "tg")
}
\arguments{
\item{ids}{A vector containing drama ids to be downloaded}

\item{includeTokens}{If set to true, the table also contains each token in an utterance}

\item{defaultCollection}{The collection prefix is added if no prefix is found}
}
\description{
Loads a CSV-formatted text from the server,
assuming the main server url has been set correctly.
}
