% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.QDUtteranceStatistics}
\alias{plot.QDUtteranceStatistics}
\title{Utterance positions}
\usage{
\method{plot}{QDUtteranceStatistics}(x, drama = NULL,
  colors = qd.colors, frame = TRUE, xlab = "Time", ...)
}
\arguments{
\item{x}{A table generated from the function}

\item{colors}{The colors to be used}

\item{...}{Parameters passed to stripchart().}

\item{segmentedText}{If supplied, act boundaries will be marked}

\item{plotFrame}{Whether to draw a frame around the plot}
}
\description{
Uses `stripchart` to  plot each utterance at their position,
in a line representing the figure. The dot is marked in the middle of each utterance.
Might look weird if very long utterances are present.
}
\seealso{
stripchart
}
