% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{report}
\alias{report}
\title{Report}
\usage{
report(id = "test:rksp.0", of = file.path(getwd(),
  paste0(unlist(strsplit(id, ":", fixed = TRUE))[2], ".html")),
  type = c("Single"), ...)
}
\arguments{
\item{id}{The id of the text or a list of ids}

\item{of}{The output file}

\item{type}{The type of the report. "Single" gives a report about a single play, 
while "Compare" can be used to compare multiple editions of a play}

\item{...}{Arguments passed through to the rmarkdown document}
}
\description{
generates a report for a specific dramatic text
}
