% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rksp.0}
\alias{rksp.0}
\title{The data set exported from Emilia Galotti}
\format{A data frame with 25365 rows and 9 columns:
\describe{
 \item{drama}{The drama id}
 \item{length}{The length of the drama (in tokens)}
 \item{begin}{The beginning of an utterance (character position)}
 \item{end}{The end of an utterance (character position)}
 \item{Speaker.figure_surface}{The surface description of a figure}
 \item{Speaker.figure_id}{}
 \item{Token.surface}{}
 \item{Token.pos}{}
 \item{Token.lemma}{}
}}
\usage{
rksp.0
}
\description{
The data set exported from Emilia Galotti
}
\keyword{datasets}

