% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{setup}
\alias{setup}
\title{This function initialises the import from XMI files.}
\usage{
setup(datadir = "/Users/reiterns/Documents/QuaDramA/Data")
}
\arguments{
\item{datadir}{A path to the directory in which data and metadata are located}
}
\description{
This function initialises the import from XMI files.
}
