% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{setup}
\alias{setup}
\title{This function initialises the import from XMI files.}
\usage{
setup(dataDirectory = file.path(path.expand("~"), "QuaDramA", "Data"))
}
\arguments{
\item{dataDirectory}{A path to the directory in which data and metadata are located. "~/QuaDramA/Data" by default}
}
\description{
This function initialises the import from XMI files.
}
