% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{setup}
\alias{setup}
\alias{setDataDirectory}
\alias{setCollectionDirectory}
\title{This function initialises the paths to data files.}
\usage{
setup(dataDirectory = file.path(path.expand("~"), "QuaDramA", "Data2"),
  collectionDirectory = file.path(dataDirectory, "collections"))

setDataDirectory(dataDirectory = file.path(path.expand("~"), "QuaDramA",
  "Data2"))


  setCollectionDirectory(collectionDirectory = file.path(getOption("qd.datadir"),
  "collections"))
}
\arguments{
\item{dataDirectory}{A path to the directory in which data and metadata are located. 
"~/QuaDramA/Data2" by default.}

\item{collectionDirectory}{A path to the directory in which collections are stored. 
By default, the directory is called "collection" below the data directory.}
}
\description{
This function initialises the paths to data files.
}
