% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{setup}
\alias{setup}
\title{This function initialises the import from XMI files.}
\usage{
setup(dataDirectory = file.path(path.expand("~"), "QuaDramA", "Data2"),
  collectionDirectory = file.path(dataDirectory, "collections"))
}
\arguments{
\item{dataDirectory}{A path to the directory in which data and metadata are located. 
"~/QuaDramA/Data" by default.}

\item{collectionDirectory}{A path to the directory in which collections are stored. 
By default, the directory is called "collection" below the data directory.}
}
\description{
This function initialises the import from XMI files.
}
