% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utteranceStatistics.R
\name{utteranceStatistics}
\alias{utteranceStatistics}
\title{Utterance Statistics}
\usage{
utteranceStatistics(drama, normalizeByDramaLength = TRUE)
}
\arguments{
\item{normalizeByDramaLength}{Logical value. If true, the resulting values will be normalized by the length of the drama.}

\item{t}{The dramatic text(s)}
}
\description{
This method calculates the length of each utterance, organised by figure and drama.
}
\examples{
data(rksp.0)
ustat <- utteranceStatistics(rksp.0)
\dontrun{
boxplot(ustat$utteranceLength ~ ustat$character,
   col=qd.colors[1:5],
   las=2, frame=FALSE)
}
}
\seealso{
\code{\link{format.QDHasCharacter}}
}
